GEODE-5985 Possible integer overflow
We had a number of calculations that were taking place with ints instead
of longs.  I've modified them to, in most places, just add an L to the
constant multiplier.  In other places I had to change a variable type
from int to long.

There were a few places that were also assigning a long to an int
variable or an int to a byte variable.  In these I added bit AND
operators to reduce the chances of overflow.

IndexElemArray needed a lot of work due to its "size" instvar being
a byte.

This closes #2783

+++ geode-core/src/main/java/org/apache/geode/admin/jmx/internal/MBeanUtil.java
@@ -303,7 +303,7 @@ public class MBeanUtil {
    * @param refreshInterval the seconds between refreshes
    */
   static void registerRefreshNotification(NotificationListener client, Object userData,
-      RefreshNotificationType type, int refreshInterval) {
+      RefreshNotificationType type, long refreshInterval) {
     if (client == null) {
       throw new IllegalArgumentException(
           "NotificationListener is required");
@@ -371,8 +371,8 @@ public class MBeanUtil {
         timerNotificationId = refreshTimer.addNotification(type.getType(), // type
             type.getMessage(), // message = "refresh"
             userData, // userData
-            new Date(System.currentTimeMillis() + refreshInterval * 1000), // first occurrence
-            refreshInterval * 1000); // period to repeat
+            new Date(System.currentTimeMillis() + refreshInterval * 1000L), // first occurrence
+            refreshInterval * 1000L); // period to repeat
 
         // put an entry into the map for the listener...
         notifications.put(type, timerNotificationId);
